// chat2db-community-client/src/typings/sqlParser.ts
export interface ISimpleTableVO {
  datasourceName: string;
  databaseName: string;
  schemaName: string;
  tableName: string;
  tableAlias: string;
  comment: string;
  insertText: string;
}

export interface ISimpleColumnVO {
  datasourceName: string;
  databaseName: string;
  schemaName: string;
  tableName: string;
  columnName: string;
  dataType: string;
  comment: string;
  insertText: string;
}
// ...
export interface SqlStatement {
  sql: string;
  sqlStartRowNum: number;
  sqlStartColNum: number;
  sqlEndRowNum: number;
  sqlEndColNum: number;
  /** The type of this SQL */
  type: SqlTypeEnum;
  /** Is this SQL valid? */
  statementType: StatementValidTypeEnum;
  /** Single line comment content before the sql statement */
  comment: string;
  /** All identifiers involved in this SQL */
  identifiers: SimpleIdentifier[];
  tableColumns: Array<{
    simpleColumns: Array<ISimpleColumnVO>;
    simpleTable: ISimpleTableVO;
  }>;
  insertValueMappings?: SimpleInsertValueMapping[];
}

export interface MarkMessage {
  endLineNum: number;
  startLineNum: number;
  startColNum: number;
  endColNum: number;
  message: string;
  /** [error, warning, ...] */
  type: string;
}

export interface SimpleIdentifier {
  name: string;
  alias: string;
  type: string;
  identifierDatabase: string;
  identifierSchema: string;
  identifierTable: string;
  identifierStartRowNum: number;
  identifierStartColNum: number;
  identifierEndColNum: number;
  identifierEndRowNum: number;
  aliasStartRowNum: number;
  aliasStartColNum: number;
  aliasEndColNum: number;
  aliasEndRowNum: number;
}

export interface SimpleInsertValueMapping {
  columnStartRowNum?: number | null;
  columnStartColNum?: number | null;
  columnEndRowNum?: number | null;
  columnEndColNum?: number | null;
  valueStartRowNum?: number | null;
  valueStartColNum?: number | null;
  valueEndRowNum?: number | null;
  valueEndColNum?: number | null;
  rowStartRowNum?: number | null;
  rowStartColNum?: number | null;
  rowEndRowNum?: number | null;
  rowEndColNum?: number | null;
  rowIndex: number;
  columnIndex: number;
  mappingStatus: InsertValueMappingStatusEnum;
}

export enum InsertValueMappingStatusEnum {
  MATCHED = 'MATCHED',
  UNMAPPED_COLUMN = 'UNMAPPED_COLUMN',
  UNMAPPED_VALUE = 'UNMAPPED_VALUE',
}
// ...
export enum StatementValidTypeEnum {
  VALID = 'VALID',
  INVALID = 'INVALID',
}

// export enum TipTypeEnum {
//   DATABASE = 'DATABASE',
//   SCHEMA = 'SCHEMA',
//   TABLE = 'TABLE',
//   VIEW = 'VIEW',
//   COLUMN = 'COLUMN',
//   PARAMETER = 'PARAMETER',
//   JOIN_CLAUSE = 'JOIN_CLAUSE',
// }

export enum SqlTypeEnum {
  SELECT = 'SELECT',
  INSERT = 'INSERT',
  UPDATE = 'UPDATE',
  DELETE = 'DELETE',
  ALTER = 'ALTER',
  CREATE = 'CREATE',
  DROP = 'DROP',
  DROP_DATABASE = 'DROP_DATABASE',
  DROP_TABLE = 'DROP_TABLE',
  DROP_VIEW = 'DROP_VIEW',
  DROP_FUNCTION = 'DROP_FUNCTION',
  DROP_PROCEDURE = 'DROP_PROCEDURE',
  DROP_USER = 'DROP_USER',
  DROP_ROLE = 'DROP_ROLE',
  DROP_EVENT = 'DROP_EVENT',
  DROP_INDEX = 'DROP_INDEX',
  DROP_TRIGGER = 'DROP_TRIGGER',
  DROP_CONSTRAINT = 'DROP_CONSTRAINT',
  DROP_SEQUENCE = 'DROP_SEQUENCE',
  DROP_COLUMN = 'DROP_COLUMN',
  DROP_SCHEMA = 'DROP_SCHEMA',
  CREATE_TABLE = 'CREATE_TABLE',
  CREATE_VIEW = 'CREATE_VIEW',
  CREATE_DATABASE = 'CREATE_DATABASE',
  CREATE_SCHEMA = 'CREATE_SCHEMA',
  CREATE_FUNCTION = 'CREATE_FUNCTION',
  CREATE_PROCEDURE = 'CREATE_PROCEDURE',
  CREATE_USER = 'CREATE_USER',
  CREATE_EVENT = 'CREATE_EVENT',
  CREATE_INDEX = 'CREATE_INDEX',
  CREATE_TRIGGER = 'CREATE_TRIGGER',
  CREATE_ROLE = 'CREATE_ROLE',
  CREATE_CONSTRAINT = 'CREATE_CONSTRAINT',
  CREATE_SEQUENCE = 'CREATE_SEQUENCE',
  CREATE_COLUMN = 'CREATE_COLUMN',
  USE_DATABASE = 'USE_DATABASE',
  SET_SCHEMA = 'SET_SCHEMA',
  DROP_ROUTINE = 'DROP_ROUTINE',
  OTHER = 'OTHER',
}